Refatorando testes.
close #92
