+++ devpro/encurtador/templates/encurtador/home.html
@@ -4,7 +4,7 @@
 {% include "encurtador/header.html" %}
 
   <body class="reduce-main text-center">
-
+    
     <div class="container">
 {% include "encurtador/titulo.html" %}
