--- devpro/encurtador/tests/test_encurtador.py
@@ -1,163 +1,0 @@
-import pytest
-from http import HTTPStatus
-from model_bakery import baker
-
-from django.test import Client
-from devpro.django_assertions import assert_contains
-from devpro.encurtador.models import UrlRedirect, UrlLog
-
-from devpro.encurtador.facade import separa_e_conta_os_redirecionamentos_por_data,\
-                                 soma_de_cliques_pela_slug, soma_de_cliques, obtem_todas_as_urls
-
-
-@pytest.fixture
-def url_encurtada(db):
-    return baker.make(UrlRedirect)
-
-
-@pytest.fixture
-def lista_urls(db):
-    return baker.make(UrlRedirect, 2)
-
-
-@pytest.fixture
-def logs(lista_urls):
-    log_url1 = baker.make(UrlLog, 3, url_redirect=lista_urls[0])
-    log_url2 = baker.make(UrlLog, 8, url_redirect=lista_urls[1])
-    return log_url1, log_url2
-
-
-def testa_a_listagem_de_todas_as_url_cadastrada(lista_urls):
-
-    listas = obtem_todas_as_urls()
-
-    # numero toral de logs da url1 acessado utilizando a chave estrangeira
-    assert len(listas) == len(UrlRedirect.objects.all())
-
-
-def testa_a_relacao_entre_o_modelo_url_redirect_e_url_log(logs, lista_urls):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa os modelos UrlRedirect e UrlLog com o ORM.
-    ---------------------------------------------------------------------------------------
-    '''
-
-    # numero total de url cadastradas
-    assert len(UrlRedirect.objects.all()) == 2
-
-    # numero total de logs
-    assert len(UrlLog.objects.all()) == len(logs[0]) + len(logs[1])
-
-    # numero toral de logs da url1 acessado utilizando a chave estrangeira
-    assert len(lista_urls[0].logs.all()) == len(logs[0])
-
-    # numero toral de logs da url2 acessado utilizando a chave estrangeira
-    assert len(lista_urls[1].logs.all()) == len(logs[1])
-
-
-def testa_a_facada_separa_e_conta_os_redirecionamentos_por_data(logs, lista_urls):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa a fachada que calcula o numero de cliques em uma url por dia
-    ---------------------------------------------------------------------------------------
-    '''
-    for url, log in zip(lista_urls, logs):
-        redirecionamentos_por_data = separa_e_conta_os_redirecionamentos_por_data(url.slug)
-        for r in redirecionamentos_por_data:
-            assert r.cliques == len(log)
-
-
-def testa_a_facada_que_calcula_o_numero_total_de_cliques_pela_slug(logs, lista_urls):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa a fachada que calcula o numero de cliques em uma url usando apenas o slug
-    ---------------------------------------------------------------------------------------
-    '''
-    for url, log in zip(lista_urls, logs):
-        assert soma_de_cliques_pela_slug(url.slug) == len(log)
-
-
-def testa_a_facada_que_calcula_o_numero_total_de_cliques(logs, lista_urls):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa a fachada que calcula o numero de cliques em uma url usando usando o
-    redirecionamentos_por_data
-    ---------------------------------------------------------------------------------------
-    '''
-    for url, log in zip(lista_urls, logs):
-        redirecionamentos_por_data = separa_e_conta_os_redirecionamentos_por_data(url.slug)
-        assert soma_de_cliques(redirecionamentos_por_data) == len(log)
-
-
-def testa_codigo_de_estato_da_home(client: Client, db):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa se o caminho para o home foi encontrado
-    ---------------------------------------------------------------------------------------
-    '''
-    response = client.get('/')
-    assert response.status_code == HTTPStatus.OK
-
-
-def testa_o_conteudo_da_home(client: Client, db):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa se o caminho para o home foi encontrado
-    ---------------------------------------------------------------------------------------
-    '''
-    response = client.get('/')
-    assert_contains(response, '<span class="color-y">Reduce</span></p>')
-
-
-def testa_redirecionamento_com_slug_existente(client: Client, url_encurtada):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa se aplicao foi redireciona para o site desejado
-    ---------------------------------------------------------------------------------------
-    '''
-    response = client.get('/' + url_encurtada.slug)
-    assert response.status_code == HTTPStatus.FOUND  # 302
-    assert response.headers['Location'] == url_encurtada.destino
-
-
-def testa_redirecionamento_com_slug_nao_existente(client: Client, db):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa se a aplicação foi rederecionada para '/' quando
-    são existe o slug
-    ---------------------------------------------------------------------------------------
-    '''
-    response = client.get('/banana')
-    assert response.status_code == HTTPStatus.FOUND  # 302
-    assert response.headers['Location'] == '/'
-
-
-def testa_codigo_de_estato_da_pagina_de_relatorio(client: Client, url_encurtada):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa o estado da pagina de relatorio
-    ---------------------------------------------------------------------------------------
-    '''
-    response = client.get(f'/relatorio/{url_encurtada.slug}')
-    assert response.status_code == HTTPStatus.OK  # 200
-
-
-def testa_se_foi_carregado_o_template_correto(client: Client, url_encurtada):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa se o template correto foi carregado
-    ---------------------------------------------------------------------------------------
-    '''
-    response = client.get(f'/relatorio/{url_encurtada.slug}')
-    assert_contains(response, '<h5>Original: <a href=')
-    assert_contains(response, '<h3>Reduzido: <a href=')
-
-
-def testa_se_url_do_destino(client: Client, url_encurtada):
-    '''
-    ---------------------------------------------------------------------------------------
-    Testa se url do destino apareceu corretament
-    ---------------------------------------------------------------------------------------
-    '''
-    response = client.get(f'/relatorio/{url_encurtada.slug}')
-    assert_contains(response, f'<a href="{url_encurtada.destino}"')
